[compatibility] curency value as float not string

+++ modules/Utils/CurrencyField/CurrencyFieldCommon_0.php
@@ -46,7 +46,7 @@ class Utils_CurrencyFieldCommon extends ModuleCommon {
 		if (!is_array($p)) $p = explode('__', $p);
                 if(!is_numeric($p[0]) && $p[0]!='') return false;
 		if (!isset($p[1])) $p[1] = Base_User_SettingsCommon::get('Utils_CurrencyField', 'default_currency');
-        $p[0] = str_replace(array(',', self::get_decimal_point($p[1])), '.', $p[0]);
+        $p[0] = (float) str_replace(array(',', self::get_decimal_point($p[1])), '.', $p[0]);
 		return $p;
 	}
